src/requirements.txt
Flask
flask_sqlalchemy
pg8000
